rename to kdf, remove thread arg
